requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.1
black==24.1.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0
django-debug-toolbar==4.2.0
Pillow==10.2.0